# Conflicts:
#	composeApp/src/commonMain/composeResources/values-ar/strings.xml
#	composeApp/src/commonMain/composeResources/values/strings.xml
#	composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/repositoryModule.kt
#	composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/useCaseModule.kt
#	composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/di/viewModelModule.kt
#	composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/exercise/CreateExerciseInteractionListener.kt
#	composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/exercise/CreateExerciseState.kt

+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/exercise/CreateExerciseInteractionListener.kt
@@ -7,7 +7,7 @@ import com.cairosquad.evolvefit.viewmodel.onboarding.models.UiImage
 interface CreateExerciseInteractionListener {
     fun onNameChanged(name: String)
     fun onImagePicked(image: UiImage)
-    fun onEquipmentToggled(equipmentId: Long)
+    fun onEquipmentToggled(equipmentId: Int)
     fun onFocusAreaNameSelected(name: String)
     fun onEquipmentNameSelected(toolName: String)
     fun onImagePickerClicked()
